implement recursive page gen
