+++ src/main.py
@@ -1,7 +1,4 @@
-import os
-import shutil
-
-from src.page_gen import generate_page
+from src import page_gen
 
 STATIC_PATH = "./static"
 PUBLIC_PATH = "./public"
@@ -11,30 +8,11 @@ TEMPLATE_PATH = "./template.html"
 
 def main():
     print("Copying static files to public directory...")
-    copy_dir(STATIC_PATH, PUBLIC_PATH)
+    page_gen.copy_dir(STATIC_PATH, PUBLIC_PATH)
 
-    print("Generating page...")
-    generate_page(
-        os.path.join(CONTENT_PATH, "index.md"),
-        TEMPLATE_PATH,
-        os.path.join(PUBLIC_PATH, "index.html"),
-    )
+    print("Generating content...")
+    page_gen.generate_pages_recursive(CONTENT_PATH, TEMPLATE_PATH, PUBLIC_PATH)
     print("Done!")
-
-
-def copy_dir(src_dir: str, dest_dir: str) -> None:
-    if os.path.exists(dest_dir):
-        shutil.rmtree(dest_dir)
-    os.mkdir(dest_dir)
-
-    for item in os.listdir(src_dir):
-        src_path = os.path.join(src_dir, item)
-        dest_path = os.path.join(dest_dir, item)
-        print(f" * {src_path} -> {dest_path}")
-        if os.path.isfile(src_path):
-            shutil.copy(src_path, dest_path)
-        else:
-            copy_dir(src_path, dest_path)
 
 
 if __name__ == "__main__":
